requirements.txt
# BigQuery AI Hackathon: Intelligent Insurance Engine
# Python dependencies for the project

# Core Google Cloud libraries
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
bigframes>=0.10.0

# Optional Google Cloud libraries (with fallback support)
google-cloud-vision>=3.4.0
google-cloud-documentai>=2.20.0

# Additional Google Cloud libraries for full functionality
google-cloud-core>=2.4.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-generativeai>=0.8.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Web interface
streamlit>=1.28.0
plotly>=5.15.0

# Image processing
Pillow>=10.0.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
faker>=19.0.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
